requirements.txt
streamlit==1.39.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
ibm-cos-sdk==2.13.6
ibm-cos-sdk-core==2.13.6
ibm-cos-sdk-s3transfer==2.13.6
python-dateutil==2.9.0.post0
ibm-watson==8.1.0
python-dotenv==1.0.1
ibm_watsonx_ai==1.0.11
numpy==1.26.4
pandas==2.1.4
ibm-cloud-sdk-core==3.20.6
Flask==3.0.3
Flask-Cors==5.0.0